Reduce duplication by embedding a fuseMount type
